task-6/requirements.txt
chromadb==1.5.0
cohere==5.20.6
groq==0.37.1
langchain-chroma==1.1.0
langchain-classic==1.0.1
langchain-cohere==0.5.0
langchain-community==0.4.1
langchain-core==1.2.13
langchain-groq==1.1.2
langchain-text-splitters==1.1.1
python-dotenv==1.2.1
